Extracting helpers into dedicated files. (#227)
* Extracting helpers into dedicated files.

* Extracting static const.
