+++ ts/helpers/const.ts
@@ -1,0 +1,4 @@
+export const ownCode = "octorelay";
+export const closeBtnId = "pop-closer";
+const closeIconHTML = '<span class="fa fa-close fa-sm"></span>';
+export const closeBtnHTML = `<button id="${closeBtnId}" type="button" class="close">${closeIconHTML}</button>`;
