+++ ts/model/initOctoRelayModel.ts
@@ -1,0 +1,8 @@
+import { OctoRelayViewModel } from "./OctoRelayModel";
+
+export const initOctoRelayModel = () => {
+  OCTOPRINT_VIEWMODELS.push({
+    construct: OctoRelayViewModel,
+    dependencies: ["settingsViewModel", "loginStateViewModel"],
+  });
+};
